Rename src dir
